+++ TrashStack.csproj
@@ -2,23 +2,29 @@
 
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
+        <LangVersion>preview</LangVersion>
     </PropertyGroup>
     <!--    <Target Name="Rollup" BeforeTargets="Build">-->
     <!--        <Exec Command="yarn build:svelte" />-->
     <!--    </Target>-->
     <ItemGroup>
-        <Watch Include="wwwroot/js/*.svelte"/>
-        <Watch Include="**\*.rest"/>
-        <Watch Include="**\*.config.json"/>
+        <Watch Include="wwwroot/js/*.svelte" />
+        <Watch Include="**\*.rest" />
+        <Watch Include="**\*.config.json" />
         <!--        match svelte js files-->
 <!--        <Watch Include="**\*.js" Exclude="node_modules\**\*;**\*.js.map;obj\**\*;bin\**\*"/>-->
     </ItemGroup>
     <ItemGroup>
-        <PackageReference Include="CodeMechanic.RazorHAT" Version="1.0.8"/>
-        <PackageReference Include="CodeMechanic.Types" Version="1.0.7"/>
-        <PackageReference Include="CodeMechanic.Regex" Version="1.0.2"/>
-        <PackageReference Include="Htmx" Version="1.7.0"/>
-        <PackageReference Include="Htmx.TagHelpers" Version="1.7.0"/>
-        <PackageReference Include="System.Runtime.Caching" Version="8.0.0"/>
+        <PackageReference Include="CodeMechanic.RazorHAT" Version="1.0.8" />
+        <PackageReference Include="CodeMechanic.Types" Version="1.0.7" />
+        <PackageReference Include="CodeMechanic.Regex" Version="1.0.2" />
+        <PackageReference Include="Htmx" Version="1.7.0" />
+        <PackageReference Include="Htmx.TagHelpers" Version="1.7.0" />
+        <PackageReference Include="System.Runtime.Caching" Version="8.0.0" />
+    </ItemGroup>
+    <ItemGroup>
+      <Reference Include="System.Runtime.Caching">
+        <HintPath>..\..\..\..\..\.nuget\packages\system.runtime.caching\8.0.0\lib\net7.0\System.Runtime.Caching.dll</HintPath>
+      </Reference>
     </ItemGroup>
 </Project>
